vers2.1

+++ src/ColoredEllipse.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 public class ColoredEllipse extends Ellipse2D.Float implements ColoredShape  {
     final Color color;
 
-    public ColoredEllipse(int x, int y, int width, int height, Color color){
+    public ColoredEllipse(float x, float y, float width, float height, Color color){
         super(x, y, width, height);
         this.color = color;
     }
